refactor(screens): creating screens folder and extracting pertinente classes to their files
