+++ my_weather/.metadata
@@ -15,22 +15,7 @@ migration:
     - platform: root
       create_revision: 300451adae589accbece3490f4396f10bdf15e6e
       base_revision: 300451adae589accbece3490f4396f10bdf15e6e
-    - platform: android
-      create_revision: 300451adae589accbece3490f4396f10bdf15e6e
-      base_revision: 300451adae589accbece3490f4396f10bdf15e6e
-    - platform: ios
-      create_revision: 300451adae589accbece3490f4396f10bdf15e6e
-      base_revision: 300451adae589accbece3490f4396f10bdf15e6e
-    - platform: linux
-      create_revision: 300451adae589accbece3490f4396f10bdf15e6e
-      base_revision: 300451adae589accbece3490f4396f10bdf15e6e
-    - platform: macos
-      create_revision: 300451adae589accbece3490f4396f10bdf15e6e
-      base_revision: 300451adae589accbece3490f4396f10bdf15e6e
     - platform: web
-      create_revision: 300451adae589accbece3490f4396f10bdf15e6e
-      base_revision: 300451adae589accbece3490f4396f10bdf15e6e
-    - platform: windows
       create_revision: 300451adae589accbece3490f4396f10bdf15e6e
       base_revision: 300451adae589accbece3490f4396f10bdf15e6e
